+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - fix bug in `windRose` where whole period is calm
 - add optimisation to `polarCluster` to speed up clustering through option `pamonce = 3`. This should not appreciably affect results.
+- fix strange bug in `aqStats` due to `lubridate` time zone issue
 
 ## openair 2.8-1